editoast: authz: define `authz::v2::roles` for organization
No code changes.
